+++ TwitterImagePipeline/Project/NSData+TIPAdditions.h
@@ -11,7 +11,8 @@
 NS_ASSUME_NONNULL_BEGIN
 
 @interface NSData (TIPAdditions)
-- (NSData *)tip_safeSubdataNoCopyWithRange:(NSRange)range;
+- (NSData *)tip_safeSubdataNoCopyWithRange:(NSRange)subRange; // throws `NSRangeException`
+- (nullable NSData *)tip_safeSubdataNoCopyWithRange:(NSRange)subRange error:(out NSError * __nullable * __nullable)outError;
 - (NSString *)tip_hexStringValue;
 @end