+++ TwitterImagePipeline/Project/TIPImageDownloadInternalContext.h
@@ -17,17 +17,24 @@ NS_ASSUME_NONNULL_BEGIN
 {
 @public
 
+    // owner
     __unsafe_unretained id<TIPImageFetchDownload> __nullable _download;
+
+    // request source state
     TIPImageDiskCacheTemporaryFile * __nullable _temporaryFile;
     TIPPartialImage * __nullable _partialImage;
     NSString * __nullable _lastModified;
     NSDictionary<NSString *, id> * __nullable _decoderConfigMap;
+
+    // internal progress state
     NSError * __nullable _progressStateError;
     NSHTTPURLResponse * __nullable _response;
     NSUInteger _contentLength;
 
+    // internal progress state flags
     struct {
         BOOL didRequestHydration:1;
+        BOOL didRequestAuthorization:1;
         BOOL didStart:1;
         BOOL didReceiveResponse:1;
         BOOL responseStatusCodeIsFailure:1;
@@ -42,10 +49,13 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (nonatomic, copy, nullable) NSURLRequest *originalRequest;
 @property (nonatomic, copy, nullable) NSURLRequest *hydratedRequest;
+@property (nonatomic, copy, nullable) NSString *authorization;
 @property (nonatomic, nullable) id<TIPImageFetchDownloadClient> client;
 @property (nonatomic, nullable) dispatch_queue_t downloadQueue;
 
 @property (nonatomic, readonly) NSUInteger delegateCount; // computed property
+
+- (void)reset; // called when retrying a download and resetting state
 
 - (NSOperationQueuePriority)downloadPriority;
 - (nullable id<TIPImageDownloadDelegate>)firstDelegate;